http protocol: close probing connection immediately
I was noticing strange behavior where an extra HTTP connection would be
open for the whole time the spider ran, which did not get any request.

This turned out to be the initial probing connection to see if port
80 is open. It is subsequentially never used until the handle goes out
of scope.

In thie patch I changed it to close the connection immediately. This saves
some resources on the server side (as well as makes single-threaded test
servers such as Python's SimpleServer not hang during probing).