Introduce Relationship Database
This commit introduces the concept of storing identifier
relationships in a seperate store from the raw crawl data.

This will make it easy to built the Analytics Framework and
identify correlations.

The future of the database structure is now partially defined
in the design/ folder, where new major design changes should
be documented.