Configurable Web Crawls
Adding the structure needed to being configurable web
crawling.

Currently there are only two options: base - which configures
the base URL (to ignore all other parts of a site and focus on
a specific set of URLs e.g. /forums) and exclude - which tells
the scanner to ignore URLs which contain one or more of the given
strings - this allows explicitly ignoring uninteresting URLs
(e.g. /profile or /settings) and also for avoiding URLs which might mess
up the scan (e.g. /logout)

This commit also fixes a bug in web crawler where the depth parameter
was overridden by a constantly updating crawl map.