OnionScan 0.2 Refactor
* New relationship database
* All old correlations now in database
* User defined relationships can now be configured
* Web UI supports searching relationships including user defined.

Apart from that, major refactor of the code to make adding new
features in the future more streamlined. OnionScan now works on a
pipeline like model, using OnionScanReport as the main interface model.

In the future, we will want to make pipelines even more modular and
ensure that we can more efficiently process larger dumps.