Add framework for testing protocol scanners
This adds a mock SOCKS5 proxy to be able to test protocol
scanners. It creates a single-threaded SOCKS5 proxy on localhost
which processes connection requests and dispatches them to a test-specific
interface, which can do further testing of the protocol specifics.

Currently this just tests available/not available for all protocols.
This is not great, but it gives surprisingly much test coverage at the
moment. This will likely go down once more protocols scanners do deeper
probing, however.